Moving where we specify the target parameter to accommodate adt 'problems'
When building a captive runtime app for the mac the order of the
parameters passed to adt is important. There's nothing in the
documentation about this, but the target parameter has to be after the
signing options otherwise adt will incorrectly complain that `Native
signing not supported on mac`.

Given how brittle art is my only worry with this commit is that it
causes problems elsewhere.